scripts/pyproject.toml
[tool.poetry]
name = "airflow2dagster-experiments"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
absl-py = "^2.1.0"
apache-airflow = "^2.9.1"
boto3 = "^1.34.89"
dagster = "^1.7.2"
dagster-aws = "^0.23.2"
dagster-dbt = "^0.23.2"
dagster-duckdb = "^0.23.2"
dagster-snowflake = "^0.23.2"
dspy-ai = "^2.4.9"
evaluate = "^0.4.2"
nltk = "^3.8.1"
langchain = "^0.1.16"
langchain-chroma = "^0.1.0"
langchain-openai = "^0.1.3"
pandas = "^2.2.2"
pydantic = "^2.7.0"
python-dotenv = "^1.0.1"
rouge-score = "^0.1.2"
typer = "^0.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
